+++ activerecord/test/cases/adapters/mysql2/schema_test.rb
@@ -11,7 +11,7 @@ module ActiveRecord
 
       def setup
         @connection = ActiveRecord::Base.connection
-        db          = Post.connection_pool.spec.underlying_configuration_hash[:database]
+        db          = Post.connection_pool.spec.db_config.configuration_hash[:database]
         table       = Post.table_name
         @db_name    = db
